requirements.txt
python-docx>=0.8.11
docxcompose>=1.4.0
pywin32>=306; sys_platform == 'win32'
